// packages/lit-analyzer/src/analyze/types/lit-target-kind.ts
export type LitTargetKind =
	| "memberFunctionElement"
	| "functionElement"
	| "constructorImplementationElement"
	| "variableElement"
	| "classElement"
	| "interfaceElement"
	| "moduleElement"
	| "memberVariableElement"
	| "constElement"
	| "enumElement"
	| "keyword"
	| "constElement"
	| "alias"
	| "moduleElement"
	| "member"
	| "label"
	| "unknown";